Cleaning up untested code and start with basic ViewSet API
I added too much code to the point it was hard to even play with and couldn't continue with development.

I'm now deleting most of the things I couldn't test to actually be able to build up functionality a bit by bit.

This PR also skips the quality checks to allow faster experimentation. Unlike unit tests, quality checks are very easy to be enabled later on.